Introducing mesh multi-resolution
- A mesh (and mesh file) now can contains many levels (max 5). When the
camera is near the object, a good resolution (level) of the mesh is
drawn and when the camera is far, a bad resolution is used.
	- Level 0 : best resolution, near the object
	- Level 1 : good resolution
	- Level 2 : medium resolution
	- Level 3 : bad resolution
	- Level 4 : worst resolution
- The notion of mesh and object radius has been introduced to treat any
object as a sphere avoiding errors during the choice of the object
level. The object radius is simply the radius of the mesh that has the
biggest radius.
- The choice of objects levels are made each 10 ticks by the simulator
by sending the LEVEL flag to the univers refresh function.
- 2 multi-resolution samples objects "Torus" and "Suzanne" have been
added. The others have been deleted.
- Some Warnings have been corrected.